requirements.txt
pandas>=2.0.0
numpy>=1.24.0
llama-cpp-python>=0.2.0
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
plotly>=5.17.0
pymysql>=1.1.0
pyyaml>=6.0
openai>=1.0.0

